+++ src/switchable/src/autorender.js
@@ -11,7 +11,7 @@ KISSY.add('switchable/autorender', function(S,DOM,JSON,Switchable) {
     Switchable.autoRender = function(hook, container) {
         hook = '.' + (hook || 'KS_Widget');
 
-        DOM.query(hook, container).each(function(elem) {
+        S.each(DOM.query(hook, container),function(elem) {
             var type = elem.getAttribute('data-widget-type'), config;
             if (type && ('Switchable Tabs Slide Carousel Accordion'.indexOf(type) > -1)) {
                 try {