+++ CHANGELOG.md
@@ -29,6 +29,7 @@ History:
  - [+] 左莫增强 calendar
  - [+] popup 增加 toggle(左莫),mouseDelay(乔花) 配置
  - [!] 去除 KISSY.XX 命名空间绑定，使用前先 use，从参数中取得模块使用
+ - [!] DOM.query 支持字符串/节点数组/单个节点，不要是KISSY或原生的NodeList
 
 ### Happy 2nd Anniversary (2011/10/26)
  - [!] Still Alive